Added test, all works
- Added test for input shape other than 30
- Moved all model tests to model.py and deleted other files
- Checked: train.py runs
- Checked: tests run when started from GUI and sample_data.pt is created
- Checked: tests run from console and sample_data.pt is created
